+++ ali/common/src/test/java/com/yanny/ali/platform/TestPlatformHelper.java
@@ -1,8 +1,11 @@
 package com.yanny.ali.platform;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
 import com.yanny.ali.api.IPlugin;
 import com.yanny.ali.platform.services.IPlatformHelper;
 import com.yanny.ali.plugin.Plugin;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.level.storage.loot.LootPool;
@@ -29,6 +32,11 @@ public class TestPlatformHelper implements IPlatformHelper {
 
     @Override
     public SpawnEggItem getSpawnEggItem(EntityType<?> entityType) {
+        return null;
+    }
+
+    @Override
+    public LootTable getLootTable(Gson gson, ResourceLocation location, JsonElement json) {
         return null;
     }
 }