+++ ali/forge/src/main/java/com/yanny/ali/forge/platform/ForgePlatformHelper.java
@@ -1,15 +1,19 @@
 package com.yanny.ali.forge.platform;
 
 import com.google.common.base.Suppliers;
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
 import com.mojang.logging.LogUtils;
 import com.yanny.ali.api.AliEntrypoint;
 import com.yanny.ali.api.IPlugin;
 import com.yanny.ali.forge.mixin.MixinLootTableForge;
 import com.yanny.ali.platform.services.IPlatformHelper;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.ForgeSpawnEggItem;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLPaths;
@@ -46,6 +50,11 @@ public class ForgePlatformHelper implements IPlatformHelper {
     @Override
     public SpawnEggItem getSpawnEggItem(EntityType<?> entityType) {
         return ForgeSpawnEggItem.fromEntityType(entityType);
+    }
+
+    @Override
+    public LootTable getLootTable(Gson gson, ResourceLocation location, JsonElement json) {
+        return ForgeHooks.loadLootTable(gson, location, json, false);
     }
 
     @NotNull