+++ ali/common/src/main/java/com/yanny/ali/platform/Services.java
@@ -1,6 +1,7 @@
 package com.yanny.ali.platform;
 
 import com.yanny.ali.platform.services.IPlatformHelper;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.ServiceLoader;
 
@@ -18,6 +19,7 @@ public class Services {
         return INSTANCE;
     }
 
+    @NotNull
     public static <T> T load(Class<T> clazz) {
         return ServiceLoader.load(clazz, clazz.getClassLoader())
                 .findFirst()