+++ ali/common/src/main/java/com/yanny/ali/platform/services/IPlatformHelper.java
@@ -1,6 +1,9 @@
 package com.yanny.ali.platform.services;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
 import com.yanny.ali.api.IPlugin;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.SpawnEggItem;
 
@@ -13,4 +16,6 @@ public interface IPlatformHelper {
         Path getConfiguration();
 
         SpawnEggItem getSpawnEggItem(EntityType<?> entityType);
+
+        LootTable getLootTable(Gson gson, ResourceLocation location, JsonElement json);
 }